Move `BaseCost`
