+++ src/moscot/costs/__init__.py
@@ -1,1 +1,1 @@
-from moscot.costs._costs import BaseCost, LeafDistance, BarcodeDistance
+from moscot.costs._costs import LeafDistance, BarcodeDistance
